+++ tests/Unit/Release/MonorepoSplitContractTest.php
@@ -19,7 +19,7 @@ class MonorepoSplitContractTest extends TestCase
         $json = $this->readJson('composer.json');
 
         $this->assertSame('toggly/feature-management-php', $json['name']);
-        $this->assertSame('1.0.0', $json['version']);
+        $this->assertSame('1.1.0', $json['version']);
         $this->assertSame(['Toggly\\FeatureManagement\\' => 'src/'], $json['autoload']['psr-4']);
         $this->assertArrayNotHasKey('Toggly\\Laravel\\', $json['autoload']['psr-4'] ?? []);
         $this->assertArrayNotHasKey('Toggly\\WordPress\\', $json['autoload']['psr-4'] ?? []);